feature/clearCheckSums: merge ReplaceChecksumReader in main reader
Merge ReplaceChecksumGraphReader into ChangelogGraphReader
